ArmPlatformPkg: make PL011UartLib baudrate casts explicit
VS2017 complains about how the UINT64-ness of BaudRate gets discarded
when calculating the UINT32 divisor for PL011 configuration.
Do the casts explicitly instead.

+++ ArmPlatformPkg/Library/PL011UartLib/PL011UartLib.c
@@ -194,7 +194,7 @@ PL011UartInitializePort (
       return RETURN_INVALID_PARAMETER;
     }
 
-    Divisor = (UartClkInHz * 4) / *BaudRate;
+    Divisor = (UINT32)((UINT64)(UartClkInHz * 4) / *BaudRate);
     Integer = Divisor >> FRACTION_PART_SIZE_IN_BITS;
     Fractional = Divisor & FRACTION_PART_MASK;
   }
